Moving admin file
